# docs/openapi/v1.yaml
openapi: 3.1.0
info:
  title: LuxeKnox Gym API
  version: 0.1.0
  summary: Shared REST contract for the Nest backend and Flutter client.
  description: |
    Source of paths: `docs/backend-frd.md` §24.
    Auth and envelope: ADR-0003.
    Prefix `/v1` lives on the server URL.

    Money is a two-decimal string. IDs are int64. Timestamps are UTC ISO-8601.
    List envelope is `{ data, meta }`. Errors are `{ code, message, details, request_id }`.

    `x-status`: `module-0` (spine), `mvp` (build with the vertical), `deferred` (ADR-0005, do not call).
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:3000/v1
    description: Local
  - url: https://api.example.com/v1
    description: Production (replace host)
...
security:
  - bearerAuth: []
paths:
...
  /employees:
    get:
      operationId: listEmployees
      summary: Staff directory
      tags: [PEOPLE]
      x-status: mvp
      x-parity: required
      x-permission: employees.read
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Q"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeePage"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "409":
          $ref: "#/components/responses/Error409"
        "422":
          $ref: "#/components/responses/Error422"
        "429":
          $ref: "#/components/responses/Error429"
    post:
      operationId: createEmployee
      summary: Create an employee
      tags: [PEOPLE]
      x-status: mvp
      x-parity: required
      x-permission: employees.create
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeCreate"
      responses:
        "201":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "409":
          $ref: "#/components/responses/Error409"
        "422":
          $ref: "#/components/responses/Error422"
        "429":
          $ref: "#/components/responses/Error429"
  /employees/{id}:
    get:
      operationId: getEmployee
      summary: Employee profile
      tags: [PEOPLE]
      x-status: mvp
      x-parity: required
      x-permission: employees.read
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "409":
          $ref: "#/components/responses/Error409"
        "422":
          $ref: "#/components/responses/Error422"
        "429":
          $ref: "#/components/responses/Error429"
    patch:
      operationId: updateEmployee
      summary: Update employee
      tags: [PEOPLE]
      x-status: mvp
      x-parity: required
      x-permission: employees.update
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "409":
          $ref: "#/components/responses/Error409"
        "422":
          $ref: "#/components/responses/Error422"
        "429":
          $ref: "#/components/responses/Error429"
  /employees/{id}/role:
    put:
      operationId: assignEmployeeRole
      summary: Assign exactly one role
      tags: [RBAC]
      x-status: mvp
      x-parity: required
      x-permission: roles.update
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignRoleRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "409":
          $ref: "#/components/responses/Error409"
        "422":
          $ref: "#/components/responses/Error422"
        "429":
          $ref: "#/components/responses/Error429"
  /employees/{id}/status:
    post:
      operationId: setEmployeeStatus
      summary: Change employment status; suspend revokes sessions
      tags: [PEOPLE]
      x-status: mvp
      x-parity: required
      x-permission: employees.update
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeStatusRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
        "409":
          $ref: "#/components/responses/Error409"
        "422":
          $ref: "#/components/responses/Error422"
        "429":
          $ref: "#/components/responses/Error429"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Access token `gk_at_<base64url>`. Never a JWT. Never a query param.
    deviceKey:
      type: apiKey
      in: header
      name: X-Device-Key
      description: Turnstile/reader credential. Scoped to attendance ingest only.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Default from gym_settings pagination.default_page_size.
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
      description: Admin tables that need page numbers.
...
    Q:
      name: q
      in: query
      schema:
        type: string
      description: Case-insensitive search (FR-API-014).
...
  responses:
    Error400:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Error401:
      description: Missing or expired session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Error403:
      description: Permission or row-scope failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Error404:
      description: Unknown or out-of-scope id
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Error409:
      description: Version mismatch or unique conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Error422:
      description: Business rule
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    Error429:
      description: Rate limited
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorCode:
      type: string
      enum:
        - validation_error
        - unauthenticated
        - forbidden
        - not_found
        - conflict
        - business_rule
        - rate_limited
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        issue:
          type: string
    ErrorBody:
      type: object
      required: [code, message, details, request_id]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"
        request_id:
          type: string
...
    Id:
      type: integer
      format: int64
      minimum: 1
    DateTime:
      type: string
      format: date-time
      description: UTC ISO-8601
    Date:
      type: string
      format: date
    PageMeta:
      type: object
      required: [limit, has_more]
      properties:
        limit:
          type: integer
        offset:
          type: integer
          nullable: true
        cursor:
          type: string
          nullable: true
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean
        total:
          type: integer
          nullable: true
          description: Present on offset lists; omitted on cursor feeds.
    UserType:
      type: string
      enum: [member, trainer, employee, admin]
    UserStatus:
      type: string
      enum: [active, inactive, suspended]
...
    User:
      type: object
      required: [id, user_type, status]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        email:
          type: string
          nullable: true
        phone_number:
          type: string
          nullable: true
        user_type:
          $ref: "#/components/schemas/UserType"
        role_id:
          $ref: "#/components/schemas/Id"
        avatar_url:
          type: string
          format: uri
          nullable: true
        status:
          $ref: "#/components/schemas/UserStatus"
        created_at:
          $ref: "#/components/schemas/DateTime"
        updated_at:
          $ref: "#/components/schemas/DateTime"
...
    EmployeeStatus:
      type: string
      enum: [active, on_probation, suspended, terminated]
    Employee:
      type: object
      required: [id, user_id, job_title, status]
      properties:
        id:
          $ref: "#/components/schemas/Id"
        user_id:
          $ref: "#/components/schemas/Id"
        job_title:
          type: string
        department:
          type: string
          nullable: true
        hire_date:
          $ref: "#/components/schemas/Date"
        status:
          $ref: "#/components/schemas/EmployeeStatus"
        role_id:
          $ref: "#/components/schemas/Id"
        user:
          $ref: "#/components/schemas/User"
    EmployeeCreate:
      type: object
      required: [email, first_name, last_name, job_title, role_id]
      properties:
        email:
          type: string
        phone_number:
          type: string
        password:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        job_title:
          type: string
        department:
          type: string
        hire_date:
          $ref: "#/components/schemas/Date"
        role_id:
          $ref: "#/components/schemas/Id"
    EmployeeUpdate:
      type: object
      properties:
        job_title:
          type: string
        department:
          type: string
        hire_date:
          $ref: "#/components/schemas/Date"
    EmployeeStatusRequest:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/EmployeeStatus"
    AssignRoleRequest:
      type: object
      required: [role_id]
      properties:
        role_id:
          $ref: "#/components/schemas/Id"
    EmployeePage:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Employee"
        meta:
          $ref: "#/components/schemas/PageMeta"
...